+++ packages/orq-rc/src/orq_ai_sdk/models/createchunkop.py
@@ -257,7 +257,7 @@ class CreateChunkResponseBody(BaseModel):
     updated_by_id: str
     r"""The unique identifier of the user who updated the chunk"""
 
-    id: Optional[str] = "chunk_01JQX6VH4453RCASNQK7637KZN"
+    id: Optional[str] = "chunk_01JQXAPP6BB08ZBQTAAJH3DEQ9"
     r"""Unique identifier for the element"""
 
     enabled: Optional[bool] = True