+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQX6VH427K1FJW5J6511XKVH"
+        "01JQXAPP69D3WZ1DW61G1FYB1V"
     )
     r"""The id of the resource"""