+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -75,7 +75,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-03T06:54:08.284Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2025-04-03T08:01:23.996Z")
     r"""The date and time the resource was created"""