+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQX6VH41E9RXTZ40A7K4CM93"
+        "01JQXAPP67RJ41JHYPE2BMS9TF"
     )
     r"""The id of the resource"""