+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -100,7 +100,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JQX6VH43CPQXTA10SNXAYWHD"
+        "01JQXAPP6A6NNAFVX15EGEZBJM"
     )
     r"""The id of the resource"""