+++ packages/orq-rc/RELEASES.md
@@ -1109,3 +1109,13 @@ Based on:
 - [python v3.3.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v3.3.0-rc.29] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.29 - packages/orq-rc
+
+## 2025-04-03 08:02:55
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.528.1 (2.565.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.3.0-rc.30] packages/orq-rc
+### Releases
+- [PyPI v3.3.0-rc.30] https://pypi.org/project/orq-ai-sdk/3.3.0-rc.30 - packages/orq-rc